# backend/static/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Within the scope of the Machinaide project, an API was developed using the Python language and Flask framework in line with the needs. This page contains information on how to use the API developed within the Machinaide project. This page explains how to send a request to the developed methods, which parameters will be sent, and what kind of results will be obtained.",
        "version": "1.0",
        "title": "Machinaide REST API Documentation",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "suat@erstesoftware.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "host": "vmi515134.contaboserver.net",
    "basePath": "/api/v1.0",
...
    "schemes": [
        "https",
        "http"
    ],
    "paths": {
...
        "/failure/getAllFailures": {
            "get": {
                "tags": [
                    "Failure"
                ],
                "summary": "This method returns a list of all fault records available in the system.",
                "description": "This method returns a list of all fault records available in the system.",
                "operationId": "getAllFailures",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "header",
                        "name": "token",
                        "description": "Tokens are required to use this method. You can get tokens using the authentication login method.",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Token"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operation successful"
                    }
                }
            }
        },
        "/failure/addFailure": {
            "post": {
                "tags": [
                    "Failure"
                ],
                "summary": "This method is used to create a new fault record.",
                "description": "This method is used to create a new fault record.",
                "operationId": "addFailure",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "header",
                        "name": "token",
                        "description": "Tokens are required to use this method. You can get tokens using the authentication login method.",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Token"
                        }
                    },
                    {
                        "in": "body",
                        "name": "payload",
                        "description": "",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Failure"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operation successful"
                    }
                }
            }
        },
        "/failure/isFailureExist": {
            "post": {
                "tags": [
                    "Failure"
                ],
                "summary": "This method is used to inquire whether the fault record, which comes as a parameter, exists in the system.",
                "description": "This method is used to inquire whether the fault record, which comes as a parameter, exists in the system.",
                "operationId": "isFailureExist",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "header",
                        "name": "token",
                        "description": "Tokens are required to use this method. You can get tokens using the authentication login method.",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Token"
                        }
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "description": "",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Failure"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operation successful"
                    },
                    "409": {
                        "description": "Data already exists"
                    }
                }
            }
        },
        "/failure/updateFailure": {
            "post": {
                "tags": [
                    "Failure"
                ],
                "summary": "This method is used to update an existing fault record.",
                "description": "This method is used to update an existing fault record.",
                "operationId": "updateFailure",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "header",
                        "name": "token",
                        "description": "Tokens are required to use this method. You can get tokens using the authentication login method.",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Token"
                        }
                    },
                    {
                        "in": "body",
                        "name": "payload",
                        "description": "",
                        "required": true,
                        "properties": {
                            "recordId": {
                                "type": "string",
                                "example": "602e6a229efcb123bdfde",
                                "description": "Is the unique ID of the fault record to be updated"
                            },
                            "sourceName": {
                                "type": "string",
                                "example": "Press030.dengeleme",
                                "description": "Shows the name of the component for which the fault record was created. [Machine].[Component].[Sensor]"
                            },
                            "factoryID": {
                                "type": "string",
                                "example": "Ermetal",
                                "description": "The unique identity of the factory you are in."
                            },
                            "sid": {
                                "type": "string",
                                "example": "dengeleme",
                                "description": "Unique ID of the part where the fault occurred."
                            },
                            "severity": {
                                "type": "string",
                                "example": "major",
                                "description": "Indicates the severity of the fault record."
                            },
                            "cost": {
                                "type": "integer",
                                "example": 1500,
                                "description": "Shows how many units the fault cost."
                            },
                            "description": {
                                "type": "string",
                                "example": "there is an oil leak.",
                                "description": "Contains an explanation about the malfunction."
                            },
                            "startTime": {
                                "type": "datetime",
                                "example": "2021-03-23T16:44",
                                "description": "Shows the date the fault started."
                            },
                            "endTime": {
                                "type": "datetime",
                                "example": "2021-03-24T16:44",
                                "description": "Shows the date the fault ended."
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operation successful"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/failure/removeFailure": {
            "delete": {
                "tags": [
                    "Failure"
                ],
                "summary": "This method is used to delete a fault record from the system.",
                "description": "This method is used to delete a fault record from the system.",
                "operationId": "removeFailure",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "header",
                        "name": "token",
                        "description": "Tokens are required to use this method. You can get tokens using the authentication login method.",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Token"
                        }
                    },
                    {
                        "in": "body",
                        "name": "payload",
                        "description": "",
                        "required": true,
                        "properties": {
                            "recordId": {
                                "type": "string",
                                "example": "602e6a229efcb123bdfde"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operation successful"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "petstore_auth": {
            "type": "oauth2",
            "authorizationUrl": "http://petstore.swagger.io/oauth/dialog",
            "flow": "implicit",
            "scopes": {
                "write:pets": "modify pets in your account",
                "read:pets": "read your pets"
            }
        },
        "api_key": {
            "type": "apiKey",
            "name": "api_key",
            "in": "header"
        }
    },
    "definitions": {
        "Token": {
            "type": "string",
            "format": "string",
            "example": ""
        },
...
        "Failure": {
            "type": "object",
            "required": [
                "factoryID",
                "sid",
                "sourceName"
            ],
            "properties": {
                "sourceName": {
                    "type": "string",
                    "example": "Press030.dengeleme",
                    "description": "Shows the name of the component for which the fault record was created. [Machine].[Component].[Sensor]"
                },
                "factoryID": {
                    "type": "string",
                    "example": "Ermetal",
                    "description": "The unique identity of the factory you are in."
                },
                "sid": {
                    "type": "string",
                    "example": "dengeleme",
                    "description": "Unique ID of the part where the fault occurred."
                },
                "severity": {
                    "type": "string",
                    "example": "major",
                    "description": "Indicates the severity of the fault record."
                },
                "cost": {
                    "type": "integer",
                    "example": 1500,
                    "description": "Shows how many units the fault cost."
                },
                "description": {
                    "type": "string",
                    "example": "there is an oil leak.",
                    "description": "Contains an explanation about the malfunction."
                },
                "startTime": {
                    "type": "datetime",
                    "example": "2021-03-23T16:44",
                    "description": "Shows the date the fault started."
                },
                "endTime": {
                    "type": "datetime",
                    "example": "2021-03-24T16:44",
                    "description": "Shows the date the fault ended."
                }
            }
        },
...
    },
...
}